backend/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.109.1"
python-dotenv = "^1.0.0"
llama-index = "^0.10.7"
llama-index-embeddings-huggingface = "^0.1.4"
llama-index-vector-stores-weaviate = "^0.1.4"
llama-index-llms-huggingface = "^0.1.3"
llama-index-llms-ollama = "^0.1.2"
llama-index-llms-openrouter = "^0.1.4"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.dependencies.llama-index-vector-stores-pinecone]
version = "^0.1.3"

[tool.poetry.dependencies.docx2txt]
version = "^0.8"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"